[backend][api] Audio to Text, Text to Summary, Text to Audio, Flask App.
This PR implements the AssemblyAPI for translating
an mp3 file to text. It also implements the CohereAPI for
translating text to a summary; CohereAPI has a max
token size of 2048 which means we need to make multiple summaries
on the text. It also implements text to image using the OpenAI API.
This PR also implements a flask app.

Known issues:
- AssemblyAPI audio -> text takes approx 10 minutes.
- CohereAPI text -> summary takes 5 minutes and isn't very good.
- OpenAI text -> best with prompt; img -> not great quality.